requirements.txt
transformers==4.53.1
requests==2.32.4
python-dotenv==1.1.1
torch==2.7.1
Flask==3.1.1
